factory, seeders

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('identification_nr/registration_nr');
+            $table->unsignedBigInteger('identification_nr/registration_nr');
             $table->unsignedInteger('type');
             $table->timestamps();
         });
