+++ database/migrations/2022_11_17_193734_create_land_properties_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('name');
-            $table->unsignedInteger('cadastral_nr');
+            $table->unsignedBigInteger('cadastral_nr');
             $table->unsignedInteger('status');
             $table->timestamps();
         });
